fix: Ensure consistent data types in mood prediction pipeline
- Updated the `EnsembleOrchestrator` to cast the PAT model to `PATModelInterface`, ensuring type safety and compatibility.
- Modified the `extract_features` method in `ProductionPATLoader` to return embeddings as `np.float32`, maintaining consistency in data types across the pipeline.
- Enhanced the model status check in the API to verify the presence of the `get_model_info` method, improving robustness.

These changes contribute to a more reliable and maintainable mood prediction system, enhancing type safety and data consistency.

+++ src/big_mood_detector/application/use_cases/predict_mood_ensemble_use_case.py
@@ -274,7 +274,7 @@ class EnsembleOrchestrator:
                 raise ValueError("No activity records provided")
 
         # Extract and return PAT embeddings (96-dim)
-        return self.pat_model.extract_features(sequence)
+        return self.pat_model.extract_features(sequence).astype(np.float64)
 
     def _predict_with_pat(
         self,
